## Why is this change being made?
There were a couple of cases where the average data could saturate.  When packages are published, the average data is reset, which prevents saturation.  However, if not publishing the data used to accumulate and saturate.  That has been resolved.  The other issue is that some fifos like tile temp have a very fast sampling rate and can saturate 16 bit, so that was changed to a 32 bit average.

Also did some data characterization, comparing different average algorithms.  There are trade-offs of computation time, accuracy for common data sets, and memory usage.
Landed on the implementation is this PR. Documented that decision with the algo.

## What is changed?
Updated 16 bit and 32 bit averaging.  Selected best for each data calculation.

## How is the change being tested?
Collected data packages on silicon.  No saturation was seen.  Decoded data was accurate for multiple scenarios.

- [ ] No test, document update only
- [x] Unit tests pass
- [x] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
This PR fixes a telemetry averaging bug by replacing the old cumulative average functions with new running average implementations that improve precision and handle saturation correctly.

#### PR Summary
The changes update the compute metrics averaging by switching from a cumulative sum method to a running summation approach with appropriate rounding and overflow protections.
- `src/services/power_tlm_mcp/data_proc_tlm_cmpnt/data_utilities.c/h`: Updated averaging functions and data structures (e.g., renaming cumulative_avg to running_avg, adding summation fields, and enhanced rounding/overflow handling).
- `src/services/power_tlm_mcp/data_proc_tlm_cmpnt/compute_metrics.c`: Modified compute metrics calculations, including renaming reset functions (to comp_metrics_reset_local_2_min_metrics) and adding in-band publishing logic.
- `src/services/power_tlm_mcp/event_trace_cmpnt/inc/telemetry_events_i.h`: Revised telemetry event definitions to align with the new running average error reporting.
- `tests/unit/services/power_tlm_mcp/data_proc_tlm_cmpnt/*`: All relevant unit tests were updated to call the new running average APIs and validate the improved averaging logic.
- Various functional test files were updated to incorporate the new `in_band_publishing_active` flag and verify correct telemetry metrics behavior.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2872430

+++ src/services/power_tlm_mcp/data_proc_tlm_cmpnt/data_sampling_i.h
@@ -29,7 +29,7 @@
 
 
 #define DOUT2MILLIVOLTS(dout) ((uint16_t)(DOUT2VOLTS((dout)) * 1000.0f))
-#define MILLIVOLTS2DOUT(mv) (VOLTS2DOUT(((mv) / 1000.0f)))
+#define MILLIVOLTS2DOUT(mv) ((uint16_t)(VOLTS2DOUT(((mv) / 1000.0f))))
 
 #define MICROSECONDS_TO_MILLISECONDS(time_diff_uS) ((time_diff_uS) / 1000)
